requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.37.1
playsound==1.3.0
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pygame==2.6.0
python-dotenv==1.0.1
simpleaudio==1.0.4
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
